refactor: 디렉토리 경로 /application/dto/stock -> /application/dto/event 변경
