+++ stock-service/src/main/java/com/destiny/stockservice/application/dto/event/cancel/StockReservationCancelEvent.java
@@ -1,4 +1,4 @@
-package com.destiny.stockservice.application.dto.stock.cancel;
+package com.destiny.stockservice.application.dto.event.cancel;
 
 import java.util.List;
 import java.util.UUID;
