+++ stock-service/src/main/java/com/destiny/stockservice/application/dto/event/reservation/StockReservationItem.java
@@ -1,4 +1,4 @@
-package com.destiny.stockservice.application.dto.stock.reservation;
+package com.destiny.stockservice.application.dto.event.reservation;
 
 import java.util.UUID;
