+++ stock-service/src/main/java/com/destiny/stockservice/application/dto/product/ProductSoldOutEvent.java
@@ -1,4 +1,4 @@
-package com.destiny.stockservice.application.dto;
+package com.destiny.stockservice.application.dto.product;
 
 import java.util.List;
 import java.util.UUID;
